issue-126 - docs: Update Documentation
