+++ docs/images/ktop-metrics-connected.png
