+++ docs/images/ktop-metrics-not-connected.png
