+++ docs/index.md
@@ -1,0 +1,7 @@
+---
+title: Home
+template: home.html
+hide:
+  - navigation
+  - toc
+---
