+++ mkdocs.yml
@@ -1,0 +1,80 @@
+site_name: ktop
+site_description: A top-like tool for Kubernetes clusters
+site_author: Vladimir Vivien
+site_url: https://vladimirvivien.github.io/ktop/
+
+repo_name: vladimirvivien/ktop
+repo_url: https://github.com/vladimirvivien/ktop
+
+theme:
+  name: material
+  custom_dir: overrides
+  logo: images/ktop.png
+  favicon: images/ktop.png
+  palette:
+    - media: "(prefers-color-scheme: light)"
+      scheme: default
+      primary: cyan
+      accent: indigo
+      toggle:
+        icon: material/brightness-7
+        name: Switch to dark mode
+    - media: "(prefers-color-scheme: dark)"
+      scheme: slate
+      primary: cyan
+      accent: indigo
+      toggle:
+        icon: material/brightness-4
+        name: Switch to light mode
+  features:
+    - navigation.tabs
+    - navigation.sections
+    - navigation.top
+    - search.highlight
+    - search.share
+    - content.code.copy
+    - content.tabs.link
+  font:
+    text: Inter
+    code: Fira Code
+
+plugins:
+  - search
+
+markdown_extensions:
+  - admonition
+  - pymdownx.details
+  - pymdownx.highlight:
+      anchor_linenums: true
+      line_spans: __span
+      pygments_lang_class: true
+  - pymdownx.inlinehilite
+  - pymdownx.snippets
+  - pymdownx.superfences
+  - pymdownx.tabbed:
+      alternate_style: true
+  - pymdownx.emoji:
+      emoji_index: !!python/name:material.extensions.emoji.twemoji
+      emoji_generator: !!python/name:material.extensions.emoji.to_svg
+  - attr_list
+  - md_in_html
+  - tables
+  - toc:
+      permalink: true
+
+extra_css:
+  - assets/stylesheets/extra.css
+
+nav:
+  - Home: index.md
+  - Guide:
+    - Getting Started: guide.md
+    - CLI Reference: cli.md
+  - Metrics:
+    - Prometheus: prometheus.md
+    - Metrics Server: metrics-server.md
+
+extra:
+  social:
+    - icon: fontawesome/brands/github
+      link: https://github.com/vladimirvivien/ktop
